% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_dictionary.R
\name{get_dictionary_table}
\alias{get_dictionary_table}
\title{Gets a dictionary table.}
\usage{
get_dictionary_table(crop)
}
\arguments{
\item{crop}{character}
}
\value{
dataframe
}
\description{
If not yet present creates a dummy one.
}
\author{
Reinhard Simon
}

