% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_dictionary.R
\name{new_dictionary_table}
\alias{new_dictionary_table}
\title{Creates an empty dictionary table}
\usage{
new_dictionary_table(crop)
}
\arguments{
\item{crop}{string}
}
\value{
dataframe
}
\description{
With dummy data
}
\author{
Reinhard Simon
}

