% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_dictionary.R
\name{post_dictionary_table}
\alias{post_dictionary_table}
\title{Posts a dictionary table locally.}
\usage{
post_dictionary_table(crop, table_dictionary)
}
\arguments{
\item{crop}{character}

\item{table_dictionary}{data.frame}
}
\description{
Posts a data.frame containing location data.
}
\author{
Reinhard Simon
}

