% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dictionary.R
\name{ui_dictionary}
\alias{ui_dictionary}
\title{ui_dictionary}
\usage{
ui_dictionary(type = "tab", title = "Dictionary configuration",
  name = "resource_dictionary", output = "hot_dictionary")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

