% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dictionary_params.R
\name{ui_dictionary_params}
\alias{ui_dictionary_params}
\title{ui_dictionary_params}
\usage{
ui_dictionary_params(name = "resource_dictionary")
}
\arguments{
\item{name}{character}
}
\description{
An interface to parameters
}
\author{
Reinhard Simon
}

